+++ R/conn_vdb.R
@@ -6,21 +6,21 @@
 #' Connecting to a VDB is required prior to calling \code{\link{makeDisplay}} or \code{\link{view}}.
 #'
 #' @param path The path on the local file system where the directory for the VDB is located
-#' 
+#'
 #' @param name A character string giving the name of the VDB. If the VDB already exists and \code{name = NULL},
 #' the previous name is used.  If the VDB exists and a string is provided for \code{name}, the name is overwritten.
 #' The primary purpose of the \code{name} argument is to facilitate deploying the trelliscope display as a Shiny app.
 #' See the \code{appName} argument of \code{\link{deployToShinyApps}}
-#' 
+#'
 #' @param autoYes should questions to proceed with directory creation operations be automatically answered with "yes"?
-#' 
+#'
 #' @param verbose should messages be printed about what is being done?
 #'
 #' @author Ryan Hafen
 #'
 #' @return An object of class \code{vdbConn} that contains the path and name of the VDB. This object is also assigned
 #' to the \code{vdbConn} option, and can be retrieved via \code{getOption("vdbConn")}
-#' 
+#'
 #' @export
 vdbConn <- function(path, name = NULL, autoYes = FALSE, verbose = TRUE) {
 
@@ -32,7 +32,7 @@ vdbConn <- function(path, name = NULL, autoYes = FALSE, verbose = TRUE) {
   stopifnot(is.character(path),
             is.logical(autoYes),
             is.logical(verbose))
-  
+
   connPath <- file.path(path, "conn.Rdata")
 
   # get VDB name, if saved
@@ -62,7 +62,7 @@ vdbConn <- function(path, name = NULL, autoYes = FALSE, verbose = TRUE) {
 
   # make sure it looks like a VDB directory
   ff <- list.files(path)
-  if(!all(c("displays", "notebook") %in% ff))
+  if(! "displays" %in% ff)
     stop(paste(path, "is not a valid VDB directory"))
 
   # Create the connection object
@@ -71,6 +71,10 @@ vdbConn <- function(path, name = NULL, autoYes = FALSE, verbose = TRUE) {
       name = name
   ), class = "vdbConn")
   save(conn, file = connPath)
+
+  # copy latest trelliscope viewer files over
+  message("*** Copying latest viewer to vdb directory...")
+  copyViewerFiles(conn)
 
   options(vdbConn = conn)
   return(conn)
@@ -89,7 +93,7 @@ print.vdbConn <- function(x, ...) {
 
 ## internal
 vdbInit <- function(path, autoYes, verbose) {
-    
+
   if(!file.exists(path)) {
     if(autoYes) {
       ans <- "y"
@@ -106,10 +110,6 @@ vdbInit <- function(path, autoYes, verbose) {
   pkgPath <- system.file(package = "trelliscope")
 
   dir.create(file.path(path, "displays"))
-
-  # if(verbose)
-  #   message("* Moving notebook files over")
-  file.copy(file.path(pkgPath, "notebook"), path, recursive = TRUE, overwrite = TRUE)
 
   TRUE
 }